pyproject.toml
[project]
name = "config"
version = "0.1.0"
description = ""
authors = [
    {name = "rodzlen",email = "qkqhrkxdkdy@naver.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "djangorestframework (>=3.15.2,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
drf-yasg = "^1.21.9"
black = "^25.1.0"
isort = "^6.0.1"
psycopg2-binary = "^2.9.10"

[tool.black]
line-length = 80

[tool.isort]
profile = "black"